Reduce the failure rate of testing total shutdown
Currently, the test is frequently failed. I found that the validators
are far behind than the best node(the observer in this test case. It
creates setup blocks), and they have been downloading the blocks when it
failed.

I think it's because of the slow synchronization. So this patch made the
validators keep alive while creating the blocks that set initial states.
The validators synchronize the chain before they shutdown.

It's not a perfect solution, but it reduces the test failure rates
dramatically.